requirements.txt
Flask
pymupdf
pytesseract
Pillow
vaderSentiment
PyPDF2

